Enhancement : Flexible diagram node shapes
uml sample, add an entity, then add a field, the field isn't in direct
edit mode. Add more fields, they'll be in direct edit mode. The fix is
to use syncExec for expression's refresh op which includes broadcasting
PartVisibilityEvent. By doing that we guarantee the gef edit part for
the shape part has been created when we try to put it in direct edit
mode.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=383924