Fix : Page book does not retain inactive pages
Page book disposes pages that are not active. This does not allow the
state of the contained parts to be retained and restored when user comes
back to the page. Examples of state that is desirable to restore include
selected tab in a tab group, weights of blocks in a split form and list
selection.

Bugzilla : https://bugs.eclipse.org/bugs/show_bug.cgi?id=414908