Fix : DataService should not allow data() method to be overridden
The data() method in DataService was initially not made final to allow
subclasses that didn't fully transition to refreshing data based on
changes in their dependencies to call refresh on every data access.

This is obviously a performance issue and we should not need this
accommodation any longer.

Bugzilla : https://bugs.eclipse.org/bugs/show_bug.cgi?id=416945