Fix : Inefficient memory utilization in XmlEditorResourceStore
XmlEditorResourceStore maintains a map of XML nodes to model elements
that should be refreshed when the XML node changes. The values in the
map are all instances of CopyOnWriteArrayList. This innefficient since
in the majority of cases, only one model element needs to be stored.
It would be better to store the element directly in the map value and
only use a list when needed.

Bugzilla : https://bugs.eclipse.org/bugs/show_bug.cgi?id=418209