Fix : Scrolling composites have incorrect background color
When background color is inherited from context, scrolling composites
render with incorrect background.

The problem can be seen in the file bug report view of EzBug sample. The
view background is supposed to be white as that's established in the
FileBugReportView class, but the view renders with gray background (on
Windows; other platforms will behave differently).

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=419550