Fix : Unable to drop onto a node if the node is itself part of a list
When performing drag-n-drop in master-details outline, one can sometimes
accomplish a drop directly onto a drop target node in order to add to
that node's children, but this feature does not work consistently and
there is no drop feedback indicating that this is possible.

The existing logic splits the drop target node vertically in half. If
the drop target is in the upper half, the insertion happens before the
drop target node. If the drop target is in the lower half, the insertion
happens after the drop target node.

An additional case is needed. If the drop target is in the middle third
and the drop target node has list factories, the drop feedback should
select the node and the insertion should be to the node's children.

The new recursive tree editor sample makes it easy to experiment with
different cases. The sample is accessible by creating a recursion.xml
file.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=420021