Fix : Confusing "none" option in required element property presentation
The radio button group presentation of the with part renderes a radio
button for each possible child type along with a "None" radio button. 
The none choice is confusing, especially when the element property is
required. This presentation is also inconsistent with the radio button
group presentation used for enum value properties.

The solution is to move the none option to the end and only show it when
the property is empty. If necessary, user can clear the property using
the provided field assist action.

A similar issue exists in the pop-up list presentation of the with part.

The solution is to align with the presentation used for value properties
with possible values, where the pop-up list selection is cleared to
indicate that the property is empty rather than using a deadicated none
list item.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=353638