Fix : JavaType validation is not refreshed after the create action
Validation for JavaType properties is not refreshed when a new type is
created by JavaTypeCreateActionHandler.

Steps to reproduce:

1. Create a java project
2. Add a file sapphire-gallery.xml
3. Go to Value properties > Java Type Names
4. In the 'Required class 1' property, type "com.test.Foo"
5. Click on the create button
6. Notice that validation continues to say that type cannot be found

The regression is due in part to turning the ValidationService into a
DataService. In the past JdtJavaTypeReferenceService called refresh on
its property when it detected a Java model change and validation would
be recomputed, but that no longer works. The solution is to broadcast an
event from JdtJavaTypeReferenceService instead. The event is caught by
ReferenceValidationService and property validation is updated.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=420348