Fix : Memory leak in DefinitionLoader
DefinitionLoader leaks a self reference when it creates a resource
loader for sdef as an anonymous inner class. The end result is that the
DefinitionLoader instance never goes out of scope, its finalize method
is not called and sdef is not disposed.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=420542