Enhancement : JavaType from an IJavaProject should expose IType
When a JavaType instance comes from an IJavaProject it wraps an IType,
but there is no API to access IType. The return type of the existing
artifact method should be modified to allow either a Class or an IType
to be returned depending on the origin of the JavaType instance.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=420816