Fix : @PossibleValues cannot be used with a list property
One can attach a PossibleValuesService implementation to either a value
property or a list property with an entry type containing a single value
property. The latter approach is required by certain property editors,
such as the slush bucket and the checkbox list. Unfortunately, the 
standard implementations of PossibleValuesService that are based on the
@PossibleValue annotation do not work with list properties.

Tests: PossibleValuesTest
Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=420907