Fix : Check box list property editor shows incorrect entry validation
When an image is attached to the list entry element, the check box list
property editor presentation displays the image with an overlay showing
the validation status. In this scenario, toggling an entry to checked
incorrectly shows a validation error if the entry's value property has
an @Required annotation.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=420926