Enhancement : SapphireWizard should not require ExecutableElement
SapphireWizard currently requires a model that extends ExecutableElement
so that execute can be invoked on wizard finish. This requirement should
be removed to make SapphireWizard more flexible. The default finish
action should check if the model is executable and execute it, but it
should be possible to do nothing on finish or to provide a custom finish
implementation when subclassing SapphireWizard.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=421077