Fix : Table cell edit not written when add is invoked
In the table property editor presentation, when a cell editor is active
with some changes and the add action is invoked via the toolbar button,
the edits are not written.

The same issue exists with move up and move down actions.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=418198