Enhancement : Node factory definition should be able to use a model path
When specifying a master-details node factory, the developer should be
able to use a model path that traverses through implied element
properties, parent navigations and/or root navigations before
terminating on an element or a list property.

This is similar to what is already possible in other constructs, such as
the property editor.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=422726