Fix : Able to finish wizard or dialog with validation errors
When the user presses on the finish button in a wizard or on the ok
button in a dialog, all delayed tasks are completed before going any
further, but there is no logic to check if the pending delayed tasks
have introduced validation errors. This makes it possible to execute
finish logic when validation errors are present, if the user is quick
enough.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=422437