Fix : Validation does not update on RelativePathService change
RelativePathValidationService does not listen to the RelativePathService
events, so there is no way to update the validation when a root changes.

Previously, this scenario could be implemented using the @DependsOn
facility, but that ceased to work when ValidationService became a
DataService.

Tests: RelativePathTests
Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=423643