Fix : Restore defaults action is missing properties
When a section contains a with part for an implied element property or
a split form, the contained properties are missing from the restore
defaults dialog. In fact, if no other parts are present in the section,
the restore defaults action is completely missing.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=423758