Fix : Unable to type some values into an editable pop-up list field
If an editable pop-up list field property editor includes possibles
values such that one value is a prefix of another, the user is unable
to type in the longer of the two values. As soon as the shorter value is
fully typed, the entered text gets selected and the next character will
replace previously typed content.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=424544