Enhancement : Label improvements for the separator part
In the sdef editor, the separator part is presented as "line separator".
The "line" qualification is not necessary as there are no other kind of
separators. Further, when a separator has a label, sdef editor should
use that label in the content outline instead of the generic "separator"
label.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=424859