Fix : Property editor is missing content after DND in the sdef editor
When a property editor is moved or copied in the sdef editor using DND,
some of definition's properties are lost. Which properties (if any) are
lost depends on which properties were present in the source definition.

One configuration that exhibits this problem is a property editor
definition consisting of a property name and an action handler.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=425886