Enhancement : Layout the container shape while direct edit node labels
While direct editing node labels, we commit the label changes thus
triggering layout on the label's container shape which increase/decrease
the available space for the cell editor. If the direct edit is canceled,
we revert to the initial value.

This change also cleans up the reconnet logic for the attached
connections when direct editing a node label. It's now in sync with
0.7.2. The reconnect logic can be removed after Konstantin's reference
service enhancement is checked in.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=436216