Fix : ProjectRelativePathService fails to relativize a path on Linux
On Linux, ProjectRelativePathService leaves a path in its absolute form
instead of converting it to its relative form.

This is a regression introduced by fixing a FindBugs warning. The
original logic for aborting the relativization process looked like this:

public IPath makeRelativeTo(Path base)
{
    //can't make relative if devices are not equal
    if (device != base.getDevice() && (device == null || !device.equalsIgnoreCase(base.getDevice())))
        return this;
    
    ...
}

The "device != base.getDevice()" portion was flagged by FindBugs for
the != comparison on strings. It looked like a misguided optimization
attempt, so I removed it. It turns out that this check was also taking
care of the case where paths do not have a device, such as on Linux. 
This is a little too clever. The replacement logic is as follows:

public Path makeRelativeTo( final Path base )
{
    // Cannot make a relative path if devices are not equal. Note that paths may not have a device on
    // some operating systems.
        
    final String baseDevice = base.getDevice();
        
    if( ( this.device != null || baseDevice != null ) &&
        ( this.device == null || ! this.device.equalsIgnoreCase( baseDevice ) ) )
    {
        return this;
    }
    
    ...
}

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=437237