Fix : Widget is disposed exception when DND a node from detached palette
In order not to cut letters when direct cell editor is invoked on a
label, we calculate the preferred size of TextFigure using SWT's Text
size. To do that, we added an SWT Text to diagram editor's figure canvas
for each TextFigure. But we didn't set it invisible. Now setting it
invisible seems to do the trick.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=437100