Fix : ScalePropertyEditorPresentation cannot be used
ScalePropertyEditorPresentation is in an internal package that's not
exported and it's not included in the set of property editors that are
queried for applicability by the framework. This makes it impossible to
use this presentation.

This regression was introduced in the 0.7 release. Prior to that
release, this presentation was located in a public package.

Rather than moving the presentation back to a public package, the fix
introduces a new style "Sapphire.PropertyEditor.Scale". The scale
presentation will activate automatically if this style is specified for
the property editor and the property is an integer property with a
@NumericRange constraint, where min and max are both specified.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=453121