Fix : Table presentation needs accessibility text
The table presentation should have the property label attached as the
accessibility text. Most times, this is not an issue, except for JAWS
reading the colon at the end of the displayed label, but if the label
is not displayed, no explanatory text is read for the field.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=458191