Fix : Default value not refreshed in the UI on Linux
When the default value changes, it is immediately refreshed in the UI
on Windows, but not on Linux where the paint events are issued somewhat
differently.

The solution is to explicitly call redraw() on the Text component when
the default value changes.

A demonstration of this scenario is available in the Gallery sample
under Value Properties -> Default Value.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=461169