Fix : Connection issue with node templates declared on same base type
When multiple diagram node templates are declared on element types that
extend from the same super type and there are embedded connections
associated with them, duplicated connection parts are created. This will
cause the connections not editable in the diagram editor. The fix is to
check the source node element type and make sure it matches the model
element type specified in the node template when creating connection
parts.

Bugzilla: https://bugs.eclipse.org/bugs/show_bug.cgi?id=463917